Fix Evaluator created for output capture.
When creating newEv from ev, shouldClose is no longer transferred, and
newEv.closePorts is now correctly called.

This problem resulted in a deadlock when ev corresponds to a form containing
output capture and not the last in the pipeline, since in that case the
output of ev is not correctly closed. In some cases, this only results in a
hanging channel but is not observable otherwise.

This fixes issue #31.