edit: Make variable completion range as small as possible.
The dollar ($), explode (@) and namespace are no longer part of the completion
range. This has a consequence in UI: Formerly, when completing a variable in a
namespace, e.g. $le:<Tab>, the candidates shown in the menu are $le:prompt,
$le:rprompt, etc. Now they are just prompt, rprompt, etc. This is only a
change in display and does not affect what is being inserted.