Introduce narrow, a generic list selection mode
Note: it is reported that this implementation has performance issue.

The purpose of narrow mode is to give user flexibility to extend list
selection related functionality by using a builtin function called
`-narrow-read`. Given the following signature, where `source` is a
function returns all the candidates; `action` is a function
which will be called on the accepted candidate.

```
narrow-read <source> <action>
```

If supported, a simple location mode can be easily implemented as:

```
bind insert C-l {
    $le:&narrow-read [({ dirs | each { echo $args[0][path] } })] {
cd $args[0] }
}
```