Add eval.ScanIntoOpts
eval.ScanIntoOpts is similar to eval.ScanOpts, instead of specifying
each option separately, it accepts an option struct and parse the
options according the struct internally.

Options organized as a group in a struct is more readable and could be
extend in the future if the calling function requires many.