Change the syntax of "use".
Previsouly, use accepts a module name a:b:x:y, and imports the file
a/b/x/y.elv (under ~/.elvish/lib) as a:b:x:y. This makes it clumsy to import
modules in deeply nested directories, because the resulting module name is
very long.

Now it is possible to omit some leading path components in the module name by
specifying them using slashes instead of colons. For example, "use a/b/c/x:y"
imports the same file a/b/c/x/y.elv, but as x:y. In other words, the part
before the last slash is used to derive the filename, but not used in the
module name.

The "use" special form used to accept a second argument for an alternative
filename. That argument has been removed to standardize the way the module
file is found. Relavite imports will be introduced some time in future to
compensate some of the lost functionality.