Support not blocking on (r)prompt beyond a timeout.
A new variable $-prompts-max-wait has been added. When the execution of
(r)prompt takes longer than this many of seconds, the editor does not wait for
it to finish; instead, it shows the last value of (r)prompt, with a stale
marker. When the (r)prompt has finishes execution, the display will be updated
again.

Use of this feature is subject to the thread unsafety of the current runtime
implementation, hence it has been marked "internal" with a leading dash.

This fixes #482.