Merge module imports into the main namespace.
A module called x is now the same as a variable called $x:, much like
how a function called f is the same as a variable called $f~.

Such module variables are not accessible yet though, due to limitations
of the current variable name resolution rules. However, those will be
fixed later.