Added simple dependency management
If an installed package has a metadata file, and if the metadata has a
`dependencies` key, then it is expected to be a list of strings
indicating packages that must also be installed.

There is still no error handling for this. If the
installation of dependencies fails, should we just uninstall everything?
Feedback welcome.