Resolve nested modules recursively.
Names like `$a:b:c:d` are resolved as if it is `$a:[b:][c:][d]`.

This also changes the resolution of variable names ending in :, so that
they can be used to refer to namespace variables.

Since namespace variables are now accessible and can be accessed like
maps, this addresses #492.