Allow for lossless json -> var -> json, and add var for adding modules
I have a usecase that involves taking JSON from a remote source,
modifying one or more fields in it, and updating the remote source
with the new JSON fields.

Elvish internally translates all the numeric fields in the recieved
JSON into strings, which causes the JSON I need to send back to be
invalid and rejected by the remote service.  Until elvish gains
support for ints and floats as native internal data types, &numberify
is a hack to turn any strings that parse as valid ints or floats into
actual ints or floats.

Elvish also translates nulls into the empty string, which does not
allow for lossless json -> var -> json translation.  To fix this, I
have added a special $nil variable to the builtin namespace that will
translate to and from a JSON null without dealing with the other
issues that using a native Go nil would.

I have also added an ExtraNS var to the runtime that will allow
for building custom shells with custom builtin namespaces.