website: Attempt to fix Makefile for GNU Make >= 4.
GNU Make >= 4 seems to behave differently with regards to dependencies
with slashes in them. In particular, the dependency on tools/* in the
wildcard rule

%.html: %.md tools/...

does not seem to cause tools/... to be built correctly. Working around
this by putting all tools in the root directory, with binaries ending in
".tool".