eval: Remove all runtime throws.
The only use of throw is in the compiler's error handling, which does
not leak into the exported API.

This fixes almost all of #570; however, part of the edit package still
uses throws for error handling now. I do not intend to migrate those
codes as they will be superseded by the rewritten package.