hashmap: Always return emptyBitmapNode when node becomes empty.
This fixes the without methods of both *bitmapNode and *collisionNode.
The former sometimes returns an empty bitmapNode that is not equal to
emptyBitmapNode (empty slice vs. nil slice). The latter returns nil,
which is plain wrong.

This manifested primary as crashes not when dissoc'ing, but when
iterating, because the iterator implementations assume that empty bitmap
nodes can only appear at the root.

This fixes elves/elvish#762.