store: Expose the test setup function, and clean up properly.
The cleanup relied on the fact that on Unix systems, files used by other
processes may be removed and the underlying filesystem resource is
released when its last reference is closed. However, this technique does
not work on Windows, leading to pollution of the temporary directory.
This commit fixes this by cleaning up the file after the testing
finishes.