cliedit/highlighter: Fix a race condition.
The race condition happens when a late update is delivered very fast, and the
lateCb will see an old state and discard the result. By locking the state mutex
for the entirety of (*Highlighter).Get, we ensure that late updates are always
delivered after the method returns.

It also serializes all calls to (*Highlighter).Get, although in practice the
calls are already serialized.