[#873] fix json output of -compileonly
The errorToJSON() method in shell/errors_to_json.go checks whether the error returned is a *parse.MultiError. However, rather than a pointer to a MultiError, an instance was being passed in. Updated the the instance methods of MultiError to take pointers as receivers to be in line with what was expected. This further required updating some other references.