pkg/cli: Simplify the LateUpdates method of Prompt and Highlighter.
The LateUpdates method of both interfaces used to return a channel of
ui.Text, but the ui.Text sent on the channel is not actually used by
App. This commit changes the method to return a channel of struct{}
instead, serving purely as a notification.

Also update other references to this method, as well as test cases.