Document and add math functions
In addition to documenting the `math:` module introduced by commit 2d8a045c
two months ago this also:

a) restructures the code to match that of the `re:` and `str:` modules, and

b) adds many more useful constants and functions.

I don't think this completely addresses issue #727 so I'm marking this
as related to, rather than fixing, that issue.

Related #727