edit: introduce filter and add-cmd-filters
Similar to `callHooks`, this change add a `callFilters` to run
filters. A filter is function which takes input and output a boolean
to indicate if the input passed or not. List of filter is chained by
logical AND.

`add-cmd-filters` will be called on current command prior to saving it
into history, which gives the user flexibility to decide whether or
not a command should be saved.