Panic during shell unit tests should be failures
While debugging my fix for issue #661 I noticed that a panic that invokes
the `rescue` function causes Go to think the test passed due to launching
a rescue shell. This introduces a mechanism for elvish unit tests to
request that panics be fatal; i.e., exit with a non-zero status rather
than launch a rescue shell.

This also changes the behavior of panics by non-interactive elvish
shells to no longer launch an unexpected rescue shell.