Proof-of-concept implementation of edit:after-prompt hook
This is an initial attempt to address #1060.

The new variable `edit:after-prompt` can contain a list of functions
that will be executed every time the prompt is redrawn. Any output
produced by the functions (either in the byte or the value stream)
must be strings, and these strings will be concatenated and added to
the output of `edit:prompt`.