Introduce replacement for package logutil
Implement an Elvish tracing package that is more flexible and easier to
use than the `logutil` package. The `trace` package allows the user more
control over the output. Such as whether the timestamps are relative
or absolute. If absolute, whether the timestamp is in the local or UTC
timezone.

This also allows for control over whether, and how much, of a Go backtrace
is included with the message and the number of frames in the backtrace at
compile time and run time. A future change will add support for including
an Elvish bactrace via a function that also takes a `eval.Frame` argument.

The new `trace` package also handles multiline messages in a more human
friendly fashion by printing each line separately. Rather than a single
line of output with newlines represented as a literal "\n" string.

Resolves #1183