Implement execution context tracing
This implements execution tracing of variable declarations, assignments,
basic commands, `if`, `for`, `while`, `try`, `and` and `or`.  There are
probably some run-time operations not logged by this change that should
be logged.  Nonetheless, this should cover all run-time ops that people
care about when doing the Elvish equivalent of `bash -x`.

There are many questions regarding the optimal execution trace output.
How much information should be included and should it be split over
multiple lines?  Should the trace message be written before or after the
operation? If written before we can't include information such as the
duration of the execution or the values returned or used as a result. If
written after then the interpretation of the trace with respect to
sub-expressions might be more difficult.

This change does not purport to answer those questions. This change
provides a concrete example of a hybrid approach that will help inform
future refinements. More importantly, this shows how increased context
is beneficial, notwithstanding the increased verbosity, compared to how
traditional POSIX shells, like Korn Shell, handle `set -o xtrace` (also,
`set -x`).

Resolves #803