pkg/parse: Introduce a new Query node type.
A Query is a sequence of compound expressions or pairs, like the arguments
and options in a command form. It is not used by any other node type, and is
intended to be a top-level node for representing a small DSL for filtering data,
such as the filter for history mode or location mode.

This addresses #1051.