Handling symlinks with an unresolvable target
Symlinks in a pathname that cannot be resolved because the target is
not valid represent a special case. Specifically, if resolving a symlink
results in a syscall.ENOENT error it is friendlier, and generally more
useful, to treat it as a non-error that simply outputs the original path.
This makes the command consistent with the behavior of the external
`realpath` it mimics.

Fixes #1236