Separate deprecation test from actual deprecations
The deprecation unit test should not depend on actual, user visible,
deprecations since they change frequently. We don't want to have to update
this test when a deprecation is retired. Too, there will hopefully be
releases with no user visible deprecations.

Also, add testing of corner cases of the deprecation level to help ensure
the mechanism works as intended.

Related #1241