Add an example of using fn created variable for function name
I was trying to add a `fn` created function to a list but was unable to
get it working. I suspected using just `name` wasn't correct and tried
`$name` and other attempts but was not successful. Finally was pointed to
`$name~` by @krader1961. I was able to find where this was documented in the
spec, but I missed it since it was just one line at the bottom of the section.
I think an example would have called out the special variable name better and
is just good to have anyway.