Build without support for the interactive daemon
This implements a mechanism for building an elvish program without
support for shared interactive history provided by the interactive
database daemon. This is done via a stub implementation of the daemon
client API. Selected at build time via:

    make ELVISH_MAKE_TAGS=elv_daemon_stub

This does not eliminate the overhead of the Bbolt DB code. That
improvement will require a different set of changes. I wanted to keep
this change small and focused on the daemon.