Replace the @elvdoc macro with a @module macro.
The @module macro is placed at the beginning of a doc, and calls elvdoc after
processing the doc.

This is needed in order to correctly put the Module anchor at the beginning of
the doc, instead of at the place that calls @elvdoc.

Also refactor the macros command to reduce boilerplate in macro implementations.