Fix coverage fluctuation.
A test in navigation_mode.go verifies that a certain key has no effect. It does
that by verifying that the buffer has not changed, which results in a race
condition: the verification might happen after the key is handled, in which
case the test works. But the verification may also happen before the key is
actually handled, in which case the key handling code has not run and thus has
no coverage.

Fix this by queueing another key that would result in an unrelated buffer
change, thus indirectly verifying that the previous key had no effect.