Improve "wrong type" error message
Produce correct exception messages caused by statements such as these:

    var x~ = $true
    ns $true

It also adds several explicit tests of the `kind-of` command. Most of
those cases are tested indirectly. Such as by the unit tests of the Go
`vals.Kind()` function. But the duplication is warranted because there
is no guarantee that `$builtin:kind-of~` behaves the same as the
`vals.Kind()` function.

Fixes #715
Fixes #744