Replace `AnyError` in tests with a specific error
The `AnyError` placeholder error can cause tests to succeed for errors
other than what was expected. That is, the use of `AnyError` can mask
bugs. So replace it with the specific error, or error type, the test
expects to be raised.

This does not remove the anyError structure because it is used in the
TestCase.DoesNotCompile() method and I want to defer updating that use
to a separate change. However, remove the public AnyError var so future
test writers don't attempt to use it.