Improve unit test output
Commit 5f1673c6 introduced the use of the go-cmp package to make it easier
to compare the wanted, versus actual, value of structures in failed unit
tests. That was very welcome.  However, it only addressed the display of
Elvish struct values. It didn't address the display of byte output, stderr
output, or exceptions (at least within the context of TestWithSetup).

For example, I was working on a change that affected TestStyledSegment.
The test ouput was this:

--- FAIL: TestStyledSegment (0.01s)
    --- FAIL: TestStyledSegment/print_(styled_(styled-segment_abc_&fg-color=cyan)_bold) (0.00s)
        builtin_fn_styled_test.go:11: got bytes out "\x1b[36;1mabc\x1b[m", want "\x1b[1;36mabc\x1b[m"

That presentation makes it unnecessarily difficult to compare the wanted
versus actual value. With this change the test output is this, which
makes it obvious why the two values aren't equal:

--- FAIL: TestStyledSegment (0.02s)
    --- FAIL: TestStyledSegment/print_(styled_(styled-segment_abc_&fg-color=cyan)_bold) (0.00s)
        builtin_fn_styled_test.go:11: bytes out:
            Wanted: "\x1b[1;36mabc\x1b[m"
            Actual: "\x1b[36;1mabc\x1b[m"

I swapped the two arguments to the cmp.Diff() call to be consistent
with the other uses of cmp.Diff(). The other uses pass the wanted and
actual values in the order I prefer.  Specifically, I find presenting the
expected value first followed by the actual value, on separate lines and
vertically aligned, makes it easier for a human to visually compare the
two values in order to understand the nature of the difference.

Changing other instances to display the wanted value before the actual
value will be done in subsequent commits. Those changes will also put
the two values on separate lines, vertically aligned, and with consistent
nomenclature.

Related #1136