Expose special paths used by Elvish via $platform:
There are several paths used by Elvish internally that should also be visible
to Elvish scripts. This change exposes three of those paths via the `platform`
module. Thus making it easier for Elvish scripts to determine where various
files live and, more importantly, should be written. The latter is especially
important for transient state files that should be placed in the same,
or a subdirectory of, the directory used for the Elvish database file.

Resolves #1385