website/cmd/ttyshot: Remove master/slave terminology.
The master/slave terminology is now increasingly shunned. The origin author of
the pty library has moved away from it back in 2019:
https://github.com/kr/pty/commit/7dc38fb350b1d71383eed149e73acb7bae231ddb

Unfortunately, renaming "master" to "pty" is problematic because it conflicts
with the package name. Use "ctrl" as a compromise - the same "ctrl" term is
already used in pkg/cli/clitest, in ttyCtrl.