website: Ignore sgr-1.
Bold texts are wider and can break alignment in ttyshots. A workaround is
possible and might come later.

This can now be noticed on ttyshots of navigation mode with default LS_COLORS,
which uses bold text for directory names, breaking the alignment of vertical
scroll bars.

This was not a problem before, since the old ttyshot code incorrectly produced
"sgr-01" instead of "sgr-1" and hid the problem. The new ttyshot code produces
sgr-1 and makes the issue apparent.