Change how `exit` works so defers are run
This replaces the direct use of os.Exit() in the `exit` implementation with
a panic so that deferred functions are run. For example, those involved
in saving profiling data.

This also adds a sanity check to the value of the `exit` command. Every
UNIX like system requires the value to be in the range [0..255]. So use
that limit for now. It is true that Windows supports the range [0..2^32)
but I don't see a good reason to support that at this time. If we find
it useful to support a different range on non-UNIX systems the check can
always be abstracted out into platform specific validation.

Fixes #1377