Move profiling back into the standard binary
This reintroduces the `-cpuprofile` flag into the standard Elvish binary. It
also adds a `-memprofile` flag. Since Elvish programs are likely to cause
lots of heap allocations that profiling option is likely to be helpful.
Adding that support has a barely measurable effect on the size of the
Elvish binary compared to just CPU profiling alone.

I measured the size of the Elvish binary with and without this change
on the platforms I use to validate the behavior of Elvish. They are all
64-bit platforms with the exception of ubuntu32 (specifically, Ubuntu
16.04 32-bit). Adding profiling increased the size of the Elvish binary
between 3.1% and 3.4%:

    macOS    3.4%
    osuse    3.1%
    fedora   3.1%
    ubuntu   3.4%
    ubuntu32 3.4%
    freebsd  3.1%
    windows  3.4%

That is obviously not a trivial increase but seems to me to be acceptable
given the probable value it adds to debugging Elvish performance issues
in the future.

Fixes #1605
Related #1377