Change `make lint` to be comprehensive
I (almost) always run `make lint` on a change I plan on turning into a
pull-request. However, the usefulness of that is limited because it only
performs two Go source code lint checks. Which is a small fraction of the
actual lint checks run in the continuous integration environment. Which
means that I am regularly surprised by CI failures due to lint checks I
didn't think to run because I expected `make lint` to run all the checks
run by the CI environment.

This change modifies the top-level "lint" target to run all the lint checks.
Also, with this change `make lint` produces no output if there is no lint.