pkg/md: Handle entities in links and autolinks correctly.
Entities in link destination and title need to be handled as part of the parsing
procedure, instead of after it. The spec doesn't have test cases for this, so
added new supplemental test cases.

Entity support in autolinks also don't have test cases in the spec, so add a
supplemental test case too.